feat: add support for ManiaPlanet4 (MP4) aka TrackMania 2.
Implemented by adding a VehicleMP4.as file and some minor
adjustments to the Dashboard.as file that is required because
CSceneVehicleVis can't be used in MP4, as such it was replaced
with CSceneVehicleVisInner. Additionally, we use the currently
viewed player's script API to get the information we want.

note that a lot of information is unavailable in MP4 that is
present in TMNEXT or ULTRA.
These are unavailable:
- Wheel rotation, rotation speed, steer angle, slip coefficient
- World velocity
- Side speed